+++ projects/core/src/main/kotlin/site/siredvin/turtlematic/data/ModRecipeProvider.kt
@@ -3,7 +3,11 @@ package site.siredvin.turtlematic.data
 import net.minecraft.data.PackOutput
 import net.minecraft.data.recipes.FinishedRecipe
 import net.minecraft.data.recipes.RecipeProvider
+import net.minecraft.tags.BlockTags
+import net.minecraft.tags.ItemTags
 import net.minecraft.world.item.crafting.Ingredient
+import net.minecraft.world.level.block.Block
+import net.minecraft.world.level.block.Blocks
 import site.siredvin.broccolium.modules.data.recipe.TweakedShapedRecipeBuilder
 import site.siredvin.broccolium.modules.data.recipe.TweakedSmithingTransformRecipeBuilder
 import site.siredvin.turtlematic.common.setup.Items
@@ -66,6 +70,14 @@ class ModRecipeProvider(output: PackOutput) : RecipeProvider(output) {
             .pattern("GDG")
             .pattern("GSG")
             .pattern("GEG")
+            .save(consumer)
+
+        TweakedShapedRecipeBuilder.shaped(Items.INSPECTION_MONOCLE.get())
+            .define('C', net.minecraft.world.item.Items.CHAIN)
+            .define('G', Blocks.GLASS)
+            .pattern(" C ")
+            .pattern("CGC")
+            .pattern(" C ")
             .save(consumer)
     }